Update README.md with creation of release with version tag only
Currently, due to the chart releaser configuration, new releases
are prefixed with the name of the helm chart. This also affects
the github action, which builds the final container image.
The current tag is currently `trivy-dojo-report-operator-0.3.5`,
and it would be much nicer, if it would just be the version number.